requirement.txt
absl-py==0.9.0
astor==0.8.1
autopep8==1.4.4
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
Django==2.0
django-redis==4.11.0
dwebsocket==0.5.12
Flask==1.1.1
gast==0.2.2
gitdb==4.0.2
githook==0.2
GitPython==3.1.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
html5lib==1.0.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kubernetes==11.0.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
smmap==3.0.1
soupsieve==2.0
sqlparse==0.3.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
uWSGI==2.0.18
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.0
wrapt==1.12.1
